requirements.txt
discord.py
aiohttp
beautifulsoup4
cryptography
python-dotenv